# backend/openapi.yaml
openapi: 3.0.0
info:
  title: 멘토-멘티 매칭 앱 API
  version: 1.0.0
  description: 멘토와 멘티를 매칭하는 앱의 API 명세
servers:
  - url: http://localhost:8080/api
    description: Development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
          enum: [mentor, mentee]
        profile:
          type: object
          properties:
            name:
              type: string
            bio:
              type: string
            imageUrl:
              type: string
            skills:
              type: array
              items:
                type: string

    MatchRequest:
      type: object
      properties:
        id:
          type: integer
        mentorId:
          type: integer
        menteeId:
          type: integer
        message:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, cancelled]

security:
  - bearerAuth: []

paths:
  /signup:
    post:
      summary: 회원가입
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - role
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
                role:
                  type: string
                  enum: [mentor, mentee]
      responses:
        "201":
          description: 회원가입 성공
        "400":
          description: 잘못된 요청
        "500":
          description: 서버 오류

  /login:
    post:
      summary: 로그인
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: 로그인 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        "400":
          description: 잘못된 요청
        "401":
          description: 인증 실패
        "500":
          description: 서버 오류

  /me:
    get:
      summary: 내 정보 조회
      responses:
        "200":
          description: 사용자 정보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: 인증 실패
        "500":
          description: 서버 오류

  /profile:
    put:
      summary: 프로필 수정
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                bio:
                  type: string
                skills:
                  type: array
                  items:
                    type: string
                image:
                  type: string
                  format: binary
      responses:
        "200":
          description: 프로필 수정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: 잘못된 요청
        "401":
          description: 인증 실패
        "500":
          description: 서버 오류

  /images/{role}/{id}:
    get:
      summary: 프로필 이미지
      parameters:
        - name: role
          in: path
          required: true
          schema:
            type: string
            enum: [mentor, mentee]
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 프로필 이미지
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        "401":
          description: 인증 실패
        "500":
          description: 서버 오류

  /mentors:
    get:
      summary: 멘토 목록 조회 (멘티 전용)
      parameters:
        - name: skill
          in: query
          schema:
            type: string
        - name: order_by
          in: query
          schema:
            type: string
            enum: [skill, name]
      responses:
        "200":
          description: 멘토 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: 인증 실패
        "500":
          description: 서버 오류

  /match-requests:
    post:
      summary: 매칭 요청 보내기 (멘티 전용)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mentorId
                - menteeId
              properties:
                mentorId:
                  type: integer
                menteeId:
                  type: integer
                message:
                  type: string
      responses:
        "200":
          description: 요청 전송 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchRequest"
        "400":
          description: 잘못된 요청
        "401":
          description: 인증 실패
        "500":
          description: 서버 오류

  /match-requests/incoming:
    get:
      summary: 받은 요청 목록 (멘토 전용)
      responses:
        "200":
          description: 받은 요청 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchRequest"
        "401":
          description: 인증 실패
        "500":
          description: 서버 오류

  /match-requests/outgoing:
    get:
      summary: 보낸 요청 목록 (멘티 전용)
      responses:
        "200":
          description: 보낸 요청 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchRequest"
        "401":
          description: 인증 실패
        "500":
          description: 서버 오류

  /match-requests/{id}/accept:
    put:
      summary: 요청 수락 (멘토 전용)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 요청 수락 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchRequest"
        "404":
          description: 요청을 찾을 수 없음
        "401":
          description: 인증 실패
        "500":
          description: 서버 오류

  /match-requests/{id}/reject:
    put:
      summary: 요청 거절 (멘토 전용)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 요청 거절 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchRequest"
        "404":
          description: 요청을 찾을 수 없음
        "401":
          description: 인증 실패
        "500":
          description: 서버 오류

  /match-requests/{id}:
    delete:
      summary: 요청 취소 (멘티 전용)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 요청 취소 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchRequest"
        "404":
          description: 요청을 찾을 수 없음
        "401":
          description: 인증 실패
        "500":
          description: 서버 오류