# backend/src/routes/requests.ts
import express, { Request, Response } from "express";
import { authenticateToken, requireRole } from "../middleware/auth";
import { run, get, all } from "../database/init";
import { MatchRequest } from "../types";

const router = express.Router();

// Send match request (mentee only)
router.post(
  "/match-requests",
  authenticateToken,
  requireRole("mentee"),
  async (req: Request, res: Response) => {
    try {
      const { mentorId, menteeId, message } = req.body;
      const userId = req.user!.id;

      if (!mentorId) {
        return res.status(400).json({ error: "Mentor ID is required" });
      }

      // Check if mentor exists and is actually a mentor
      const mentor = await get(
        'SELECT * FROM users WHERE id = ? AND role = "mentor"',
        [mentorId]
      );
      if (!mentor) {
        return res.status(400).json({ error: "Mentor not found" });
      }

      // Check if request already exists
      const existingRequest = await get(
        "SELECT * FROM match_requests WHERE mentor_id = ? AND mentee_id = ?",
        [mentorId, userId]
      );
      if (existingRequest) {
        return res.status(400).json({ error: "Request already exists" });
      }

      // Check if mentee has pending request
      const pendingRequest = await get(
        'SELECT * FROM match_requests WHERE mentee_id = ? AND status = "pending"',
        [userId]
      );
      if (pendingRequest) {
        return res
          .status(400)
          .json({ error: "You already have a pending request" });
      }

      // Create match request
      const result = await run(
        'INSERT INTO match_requests (mentor_id, mentee_id, message, status) VALUES (?, ?, ?, "pending")',
        [mentorId, userId, message || null]
      );

      const newRequest = (await get(
        "SELECT * FROM match_requests WHERE id = ?",
        [result.lastID]
      )) as MatchRequest;

      res.json({
        id: newRequest.id,
        mentorId: newRequest.mentor_id,
        menteeId: newRequest.mentee_id,
        message: newRequest.message,
        status: newRequest.status,
      });
    } catch (error) {
      console.error("Send match request error:", error);
      res.status(500).json({ error: "Internal server error" });
    }
  }
);

// Get incoming requests (mentor only)
router.get(
  "/match-requests/incoming",
  authenticateToken,
  requireRole("mentor"),
  async (req: Request, res: Response) => {
    try {
      const userId = req.user!.id;

      const requests = (await all(
        "SELECT * FROM match_requests WHERE mentor_id = ? ORDER BY created_at DESC",
        [userId]
      )) as MatchRequest[];

      const requestList = requests.map((request) => ({
        id: request.id,
        mentorId: request.mentor_id,
        menteeId: request.mentee_id,
        message: request.message,
        status: request.status,
      }));

      res.json(requestList);
    } catch (error) {
      console.error("Get incoming requests error:", error);
      res.status(500).json({ error: "Internal server error" });
    }
  }
);

// Get outgoing requests (mentee only)
router.get(
  "/match-requests/outgoing",
  authenticateToken,
  requireRole("mentee"),
  async (req: Request, res: Response) => {
    try {
      const userId = req.user!.id;

      const requests = (await all(
        "SELECT * FROM match_requests WHERE mentee_id = ? ORDER BY created_at DESC",
        [userId]
      )) as MatchRequest[];

      const requestList = requests.map((request) => ({
        id: request.id,
        mentorId: request.mentor_id,
        menteeId: request.mentee_id,
        status: request.status,
      }));

      res.json(requestList);
    } catch (error) {
      console.error("Get outgoing requests error:", error);
      res.status(500).json({ error: "Internal server error" });
    }
  }
);

// Accept request (mentor only)
router.put(
  "/match-requests/:id/accept",
  authenticateToken,
  requireRole("mentor"),
  async (req: Request, res: Response) => {
    try {
      const requestId = req.params.id;
      const userId = req.user!.id;

      // Get the request
      const request = (await get(
        "SELECT * FROM match_requests WHERE id = ? AND mentor_id = ?",
        [requestId, userId]
      )) as MatchRequest;

      if (!request) {
        return res.status(404).json({ error: "Request not found" });
      }

      // Check if mentor already has accepted request
      const acceptedRequest = await get(
        'SELECT * FROM match_requests WHERE mentor_id = ? AND status = "accepted"',
        [userId]
      );
      if (acceptedRequest && acceptedRequest.id !== parseInt(requestId)) {
        return res
          .status(400)
          .json({ error: "You already have an accepted request" });
      }

      // Accept the request
      await run(
        'UPDATE match_requests SET status = "accepted", updated_at = CURRENT_TIMESTAMP WHERE id = ?',
        [requestId]
      );

      // Reject all other pending requests for this mentor
      await run(
        'UPDATE match_requests SET status = "rejected", updated_at = CURRENT_TIMESTAMP WHERE mentor_id = ? AND id != ? AND status = "pending"',
        [userId, requestId]
      );

      const updatedRequest = (await get(
        "SELECT * FROM match_requests WHERE id = ?",
        [requestId]
      )) as MatchRequest;

      res.json({
        id: updatedRequest.id,
        mentorId: updatedRequest.mentor_id,
        menteeId: updatedRequest.mentee_id,
        message: updatedRequest.message,
        status: updatedRequest.status,
      });
    } catch (error) {
      console.error("Accept request error:", error);
      res.status(500).json({ error: "Internal server error" });
    }
  }
);

// Reject request (mentor only)
router.put(
  "/match-requests/:id/reject",
  authenticateToken,
  requireRole("mentor"),
  async (req: Request, res: Response) => {
    try {
      const requestId = req.params.id;
      const userId = req.user!.id;

      // Get the request
      const request = (await get(
        "SELECT * FROM match_requests WHERE id = ? AND mentor_id = ?",
        [requestId, userId]
      )) as MatchRequest;

      if (!request) {
        return res.status(404).json({ error: "Request not found" });
      }

      // Reject the request
      await run(
        'UPDATE match_requests SET status = "rejected", updated_at = CURRENT_TIMESTAMP WHERE id = ?',
        [requestId]
      );

      const updatedRequest = (await get(
        "SELECT * FROM match_requests WHERE id = ?",
        [requestId]
      )) as MatchRequest;

      res.json({
        id: updatedRequest.id,
        mentorId: updatedRequest.mentor_id,
        menteeId: updatedRequest.mentee_id,
        message: updatedRequest.message,
        status: updatedRequest.status,
      });
    } catch (error) {
      console.error("Reject request error:", error);
      res.status(500).json({ error: "Internal server error" });
    }
  }
);

// Cancel request (mentee only)
router.delete(
  "/match-requests/:id",
  authenticateToken,
  requireRole("mentee"),
  async (req: Request, res: Response) => {
    try {
      const requestId = req.params.id;
      const userId = req.user!.id;

      // Get the request
      const request = (await get(
        "SELECT * FROM match_requests WHERE id = ? AND mentee_id = ?",
        [requestId, userId]
      )) as MatchRequest;

      if (!request) {
        return res.status(404).json({ error: "Request not found" });
      }

      // Cancel the request
      await run(
        'UPDATE match_requests SET status = "cancelled", updated_at = CURRENT_TIMESTAMP WHERE id = ?',
        [requestId]
      );

      const updatedRequest = (await get(
        "SELECT * FROM match_requests WHERE id = ?",
        [requestId]
      )) as MatchRequest;

      res.json({
        id: updatedRequest.id,
        mentorId: updatedRequest.mentor_id,
        menteeId: updatedRequest.mentee_id,
        message: updatedRequest.message,
        status: updatedRequest.status,
      });
    } catch (error) {
      console.error("Cancel request error:", error);
      res.status(500).json({ error: "Internal server error" });
    }
  }
);

export default router;